No more OCEAN_VERSION
The OCEAN_VERSION environment variable only occurred in the start_ocean.sh script (and in no other Barge files).

It was set to stable by default but that never got used by anything. If the `--latest` option occurs, then it got changed to `latest` but we don't need to use an environment variable for that.